+++ test/cldr_test.exs
@@ -222,6 +222,6 @@ defmodule Cldr.Test do
   end
 
   test "that we get the correct default json library" do
-    assert Cldr.Config.json_library == Jason
+    assert Cldr.Config.json_library() == Jason
   end
 end